+++ src/axum.rs
@@ -13,7 +13,6 @@
 //! # #[async_trait]
 //! # impl ezsockets::SessionExt for MySession {
 //! #   type ID = u16;
-//! #   type Args = ();
 //! #   type Call = ();
 //! #   fn id(&self) -> &Self::ID { unimplemented!() }
 //! #   async fn on_text(&mut self, text: String) -> Result<(), ezsockets::Error> { unimplemented!() }
@@ -28,7 +27,7 @@
 //!     // ...
 //!    # type Session = MySession;
 //!    # type Call = ();
-//!    # async fn on_connect(&mut self, socket: ezsockets::Socket, address: std::net::SocketAddr, _args: ()) -> Result<ezsockets::Session<u16, ()>, ezsockets::Error> { unimplemented!() }
+//!    # async fn on_connect(&mut self, socket: ezsockets::Socket, request: ezsockets::Request, address: std::net::SocketAddr) -> Result<ezsockets::Session<u16, ()>, ezsockets::Error> { unimplemented!() }
 //!    # async fn on_disconnect(&mut self, id: <Self::Session as ezsockets::SessionExt>::ID) -> Result<(), ezsockets::Error> { unimplemented!() }
 //!    # async fn on_call(&mut self, call: Self::Call) -> Result<(), ezsockets::Error> { unimplemented!() }
 //! }
@@ -56,19 +55,17 @@
 //!     Extension(server): Extension<ezsockets::Server<EchoServer>>,
 //!     ezsocket: Upgrade,
 //! ) -> impl IntoResponse {
-//!     ezsocket.on_upgrade(server, ())
+//!     ezsocket.on_upgrade(server)
 //! }
 //! ```
 
 use axum_crate as axum;
-use http::Request;
 
 use crate::CloseCode;
 use crate::CloseFrame;
 use crate::RawMessage;
 use crate::Server;
 use crate::ServerExt;
-use crate::SessionExt;
 use crate::Socket;
 use async_trait::async_trait;
 use axum::extract::ws;
@@ -89,6 +86,7 @@ use std::net::SocketAddr;
 pub struct Upgrade {
     ws: ws::WebSocketUpgrade,
     address: SocketAddr,
+    request: crate::Request,
 }
 
 #[async_trait]
@@ -99,15 +97,26 @@ where
 {
     type Rejection = WebSocketUpgradeRejection;
 
-    async fn from_request(req: Request<B>, state: &S) -> Result<Self, Self::Rejection> {
+    async fn from_request(req: http::Request<B>, state: &S) -> Result<Self, Self::Rejection> {
         let ConnectInfo(address) = req
             .extensions()
             .get::<ConnectInfo<SocketAddr>>()
             .expect("Axum Server must be created with `axum::Router::into_make_service_with_connect_info::<SocketAddr, _>()`")
             .to_owned();
+
+        let mut pure_req = crate::Request::builder()
+            .method(req.method())
+            .uri(req.uri())
+            .version(req.version());
+        for (k, v) in req.headers() {
+            pure_req = pure_req.header(k, v);
+        }
+        let pure_req = pure_req.body(()).unwrap();
+
         Ok(Self {
             ws: ws::WebSocketUpgrade::from_request(req, state).await?,
             address,
+            request: pure_req,
         })
     }
 }
@@ -151,14 +160,10 @@ impl Upgrade {
     /// When using `WebSocketUpgrade`, the response produced by this method
     /// should be returned from the handler. See the [module docs](self) for an
     /// example.
-    pub fn on_upgrade<E: ServerExt + 'static>(
-        self,
-        server: Server<E>,
-        args: <E::Session as SessionExt>::Args,
-    ) -> Response {
+    pub fn on_upgrade<E: ServerExt + 'static>(self, server: Server<E>) -> Response {
         self.ws.on_upgrade(move |socket| async move {
             let socket = Socket::new(socket, Default::default()); // TODO: Make it really configurable via Extensions
-            server.accept(socket, self.address, args).await;
+            server.accept(socket, self.request, self.address).await;
         })
     }
 }