+++ tests/chat.rs
@@ -1,5 +1,6 @@
 use async_trait::async_trait;
 use ezsockets::Error;
+use ezsockets::Request;
 use ezsockets::Server;
 use ezsockets::Socket;
 use std::collections::HashMap;
@@ -49,9 +50,11 @@ impl ezsockets::ServerExt for ChatServer {
     async fn on_connect(
         &mut self,
         socket: Socket,
+        request: Request,
         _address: SocketAddr,
-        _args: <Self::Session as ezsockets::SessionExt>::Args,
     ) -> Result<Session, Error> {
+        let value = request.headers().get("Some-Header").unwrap();
+        assert_eq!(value, "someValue");
         let id = (0..).find(|i| !self.sessions.contains_key(i)).unwrap_or(0);
         let session = Session::create(
             |_handle| SessionActor {
@@ -151,7 +154,6 @@ pub struct SessionActor {
 #[async_trait]
 impl ezsockets::SessionExt for SessionActor {
     type ID = SessionID;
-    type Args = ();
     type Call = ();
 
     fn id(&self) -> &Self::ID {